backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
httpx==0.26.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
google-genai>=1.0.0
